+++ MyERP.Web/Models/PurchaseInvoiceLineViewModels.cs
@@ -137,7 +137,7 @@ namespace MyERP.Web.Models
 
         [Display(Name = "Vat %")]
         [DisplayFormat(DataFormatString = "{0}")]
-        public Decimal VatPercentage { get; set; }
+        public Int16 VatPercentage { get; set; }
 
         [Display(Name = "Vat Amount")]
         [DisplayFormat(DataFormatString = "{0:n2}")]
